third_party/WebKit/LayoutTests/virtual/color_space/fast/canvas/color-space/canvas-drawImage-rec2020.html

<!DOCTYPE HTML>
<script src="../../../../../resources/testharness.js"></script>
<script src="../../../../../resources/testharnessreport.js"></script>
<script>

var rec2020Red = [0.627, 0.069, 0.016, 1];
var rec2020Green = [0.329, 0.919, 0.088, 1];
var rec2020Blue = [0.043, 0.011, 0.895, 1];
var rec2020Black = [0, 0, 0, 1];

function testPixels(ctx, tests)
{
    var actual, expected, tolerance = 0.01;
    for (var i = 3; i < tests.length; i++) {
      actual = ctx.getImageData(tests[i][0], tests[i][1], 1, 1).dataUnion;
      expected = tests[i][2];
      assert_true(actual.length === expected.length);
      for (var i = 0; i < actual.length; i++)
        assert_approx_equals(actual[i], expected[i], tolerance);
    }
}

function drawSRGBImageOnP3Canvas(source)
{
  var canvas = document.createElement('canvas');
  canvas.width = 20;
  canvas.height = 20;
  var ctx = canvas.getContext('2d',
      {colorSpace: 'rec2020', pixelFormat:'float16'});
  ctx.drawImage(source, 0, 0);
  var tests = [[5, 5, rec2020Red], 
               [15, 5, rec2020Green], 
               [5, 15, rec2020Blue], 
               [15, 15, rec2020Black]];
  testPixels(ctx, tests);
}

promise_test(function() {
  return new Promise((resolve, reject) => {
    var image = new Image();
    image.onload = function() {
      resolve(image);
    }
    image.src = 'resources/pattern-srgb.png'
  }).then(drawSRGBImageOnP3Canvas);
}, 'Draw SRGB image on a Rec2020 canvas and read back the Rec2020 pixels.');

// TODO(zakerinasab): crbug.com/713867
// Add at least two more tests to load pattern-p3d65.png and pattern-rec2020.png
// and draw them on a Rec2020 canvas.

</script>
